// types/contentful/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    DateTime: any;
    Dimension: any;
    HexColor: any;
    JSON: any;
    Quality: any;
};

/** Represents a binary file in a space. An asset can be any file type. */
export type Asset = {
    __typename?: 'Asset';
    contentType?: Maybe<Scalars['String']>;
    contentfulMetadata: ContentfulMetadata;
    description?: Maybe<Scalars['String']>;
    fileName?: Maybe<Scalars['String']>;
    height?: Maybe<Scalars['Int']>;
    linkedFrom?: Maybe<AssetLinkingCollections>;
    size?: Maybe<Scalars['Int']>;
    sys: Sys;
    title?: Maybe<Scalars['String']>;
    url?: Maybe<Scalars['String']>;
    width?: Maybe<Scalars['Int']>;
};
// ...
export type AssetLinkingCollections = {
    __typename?: 'AssetLinkingCollections';
    componentAuthorCollection?: Maybe<ComponentAuthorCollection>;
    componentRichImageCollection?: Maybe<ComponentRichImageCollection>;
    componentSeoCollection?: Maybe<ComponentSeoCollection>;
    entryCollection?: Maybe<EntryCollection>;
    pageBlogPostCollection?: Maybe<PageBlogPostCollection>;
};
// ...
export type ComponentRichImage = Entry & _Node & {
    __typename?: 'ComponentRichImage';
    _id: Scalars['ID'];
    caption?: Maybe<Scalars['String']>;
    contentfulMetadata: ContentfulMetadata;
    fullWidth?: Maybe<Scalars['Boolean']>;
    image?: Maybe<Asset>;
    internalName?: Maybe<Scalars['String']>;
    linkedFrom?: Maybe<ComponentRichImageLinkingCollections>;
    sys: Sys;
};
// ...
export type ComponentRichImageLinkingCollections = {
    __typename?: 'ComponentRichImageLinkingCollections';
    entryCollection?: Maybe<EntryCollection>;
};
// ...
export type ContentfulMetadata = {
    __typename?: 'ContentfulMetadata';
    tags: Array<Maybe<ContentfulTag>>;
};
// ...
export type ContentfulTag = {
    __typename?: 'ContentfulTag';
    id?: Maybe<Scalars['String']>;
    name?: Maybe<Scalars['String']>;
};

export type Entry = {
    contentfulMetadata: ContentfulMetadata;
    sys: Sys;
};

export type EntryCollection = {
    __typename?: 'EntryCollection';
    items: Array<Maybe<Entry>>;
    limit: Scalars['Int'];
    skip: Scalars['Int'];
    total: Scalars['Int'];
};
// ...
export type Sys = {
    __typename?: 'Sys';
    environmentId: Scalars['String'];
    firstPublishedAt?: Maybe<Scalars['DateTime']>;
    id: Scalars['String'];
    /** The locale that was requested. */
    locale?: Maybe<Scalars['String']>;
    publishedAt?: Maybe<Scalars['DateTime']>;
    publishedVersion?: Maybe<Scalars['Int']>;
    spaceId: Scalars['String'];
};
// ...
export type _Node = {
    _id: Scalars['ID'];
};